Fixed maximal mode
Now the maximal mode is functioning correctly. Here is a description of
the problems:

The problem was as follows:

The maximal mining mode behaves badly in 2 ways. Number 1, if there are
parallel threads it doesn't consistently report the same sets. However,
in non-parallel mode it does. Conversely the size of the set filtered
(and its contents by short label) is the same for both parallel and
non-parallel. Number 2, it is filtering some non-maximal examples but
not all of them.

The answer to both problems is actually the same. It has to do with the
computation of `nonOverlapping`. It chooses one (and only one) parent.
However, we really need all of the parents to use the current maximal
strategy.

The fix is actually simple (if potentially computationally intensive).
Right now we do not recalculate the parent for a graph. Rather we pass
the actual parent forward with each graph for later filtering. This is
nice because it means we never have to recalculate those graphs.

The problem is that do to `nonOverlapping` we are missing parents. We
could pass them forward (which could get big turning off
`nonOverlapping` leads to a huge explosion of work). So the solution:
compute all possible parents when filtering. This will require a
modification to goiso.SubGraph to allow shrinking a graph 1 edge (or
vertex) at a time.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>