Add buffer to collector channels (go 1.5 fix)
This commit stops the errors seen in

- https://github.com/golang/go/issues/11035
- specifically https://github.com/golang/go/issues/11035#issuecomment-108481131

However, in general it is not ideal to make these channels buffered and
really they should not need to be buffered. Past experience has shown
that buffering the "send to collector" channel simply masks performance
problems until sufficient memory pressure is reached. It is far better
to distribute the work evenly accross all the collectors (which was
implemented in 05fa9e9db3959de0f9cc790f416dedaed2f41b5d

If the Go 1.5 runtime bug gets fixed this change can be safely reverted.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>