Use sumpow instead of inverse to compute N
N is the fundemental matrix of the absorbing markov chain. This commit
tested out if doing this would be faster. I do not believe it is.
Further analysis shows that this is likely O(N^4) where as doing the
inversion is O(N^3). O(N^3) is not great when the matrix is 10000 + rows
which can often be for larger matrices. We need to explore faster
inversion strategies.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>