Improved the speed of selection_probability
This commit improves the speed by approximate 4x for large matrices (eg.
the only ones that matter). The difference is the inverse is less
accurate because I am using a function `sparse.spilu` which computes an
approximation of the LU decomposition. I then use that to solve for the
inverse.

There may be further gains to be made if I follow up on the docs here:
http://docs.scipy.org/doc/scipy-dev/reference/generated/scipy.sparse.linalg.use_solver.html#scipy.sparse.linalg.use_solver

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>