Computing conditional probabilities
The conditionals program computes the conditional probability of a
subgraph being selected given all of the selection probabilities and the
number of attempts to select all of the subgraphs.

So lets say you were collect a sample size of 100. You end up with 50
unique subgraphs. It tool 1000 tries to get that sample. You would
configure conditionals like so:

    conditionals -s 1000 -b /path/to/output

The number of tries is now stored at

    /path/to/output/tries

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>